Decline a German noun

Note that this function is far from covering all grammatical possibilities. I.a. it currently doesn't support the special cases where a masculine noun gets an n or en ending attached in the accusative, dative and genitive cases.

Usage

decline_noun_de(
  noun,
  gender = c("feminine", "masculine", "neuter"),
  case = c("nominative", "genitive", "dative", "accusative"),
  is_plural = FALSE
)

Arguments

noun

The noun to decline, optionally with preceding adjective(s), in (indefinite) nominative. A character scalar.

gender

The grammatical gender of the noun.

case

The grammatical case of the noun.

is_plural

Whether or not noun is a plural.

Value

A character scalar.

See also