Decline the German definite article

The declination of the German definite article is a relative pronoun.

Usage

definite_article_de_declined(
  gender = c("feminine", "masculine", "neuter"),
  case = c("nominative", "genitive", "dative", "accusative"),
  pluralize = FALSE
)

Arguments

gender

The grammatical gender of the subject the definite article refers to.

case

The grammatical case of the subject the definite article refers to.

pluralize

Whether to return the plural form of the definite article. A logical scalar.

Value

A character scalar.

Details

Note that the second singular feminine and plural genitive form "derer" is ignored by this function.

See also