Download Pandoc release

Downloads the assets and extracts the executable program binaries for Linux, macOS and Windows of a certain Pandoc release.

Usage

download_pandoc_binaries(
  release_id = pandoc_release_id_latest(),
  os = c("linux", "macos", "windows"),
  path = "bin/",
  overwrite = TRUE
)

Arguments

release_id

The GitHub release ID of the desired Pandoc release. Use pandoc_releases() to determine the release ID of a specific Pandoc version number. An integer scalar.

os

The operating system(s) for which Pandoc binaries should be downloaded. Any combination of "linux","macos" and "windows".

path

The filesystem path to which the binaries are saved to. A path or something coercible to.

overwrite

Whether to overwrite existing binaries under path.

Value

path invisibly.

See also