Generate Quarto knitr code chunks from structured data

Generates Quarto knitr figure or table code chunks from structured data.

Usage

gen_quarto_snips(data, path, type = c("fig", "tbl"))

Arguments

data

Input data as returned by read_quarto_chunk_data(). A dataframe where column names correspond to chunk options supported by quarto_fig_chunk() or quarto_tbl_chunk(), depending on type. Columns body, label and fig_cap/tbl_cap are mandatory.

path

Output directory where the generated figure chunk .qmd files are to be saved.

type

Quarto code chunk type. One of "fig" or "tbl".

Value

data, invisibly.

Details

The column names of data can either correspond to quarto_fig_chunk()/quarto_tbl_chunk()'s argument names (in snake case), to knitr's figure chunk option names (period-separared) or directly to Quarto's figure/table code chunk option names (in kebab case), they will be converted as needed. Additional column names are silently ignored.

The generated .qmd files are intended to be included in other Quarto documents via the built-in include shortcode.

See also

Examples

tmp_path <- fs::path_temp("example-gen_quarto_snips") |> fs::dir_create()

tibble::tibble(label = "fig-cars",
               fig_cap = "Today's ({lubridate::today()}) top figure",
               fig_column = "page",
               fig_height = 8L,
               body = "plot(cars)") |>
  tibble::add_row(label = "fig-mtcars",
                  fig_cap = "More cars",
                  fig_column = "screen",
                  fig_height = 8L,
                  body = "plot(mtcars)") |>
  salim::gen_quarto_snips(type = "fig",
                          path = tmp_path)

fs::dir_delete(tmp_path)