List Pandoc release assets

Uses gh::gh() to fetch filenames, corresponding operating systems and download URLs of a specific GitHub release of Pandoc via GitHub's REST API and returns them as a tibble.

Usage

pandoc_release_assets(
  release_id = yay::gh_release_latest(owner = "jgm", name = "pandoc")$id
)

Arguments

release_id

The GitHub release identifier of the desired Pandoc release. Use yay::gh_releases(owner = "jgm", name = "pandoc") to determine the release identifier of a specific Pandoc version number. An integer scalar.

Value

A tibble.

See also