Convert an integer into spelled abbreviated English or German rank

Convert an integer into spelled abbreviated English or German rank

Usage

rank_nr(x, lang = c("en", "de"))

Arguments

x

An integer scalar or something coercible to.

lang

The language to write the rank in.

Value

A character scalar.

See also