Update R packages from rpkg.dev

Installs/updates R packages from rpkg.dev to the latest (development) version.

Usage

update_rpkgs(
  pkgs = c("gitlab", "nocodb", "pal", "pkgdown.tpl", "pkgpins", "pkgpurl", "pkgsnip",
    "plotlee", "qstnr", "quappo", "rstd", "swissmuni", "tocr", "yay"),
  dependencies = TRUE,
  from_cran = FALSE
)

Arguments

pkgs

R pkgs to be updated. A subset of:

"gitlab"

"nocodb"

"pal"

"pkgdown.tpl"

"pkgpins"

"pkgpurl"

"pkgsnip"

"plotlee"

"qstnr"

"quappo"

"rstd"

"swissmuni"

"tocr"

"yay"

dependencies

Whether or not to also install uninstalled packages which pkgs depend on/link to/import/suggest (and so on recursively). See install.packages() for details.

from_cran

Whether or not to install pkgs from a CRAN repository if possible. The repository URLs set in the repos R option are used by default, with fallback to https://cloud.r-project.org/. If FALSE, the latest development versions are always installed, regardless of whether pkgs are available on CRAN or not.

Value

pkgs, invisibly.

See also