Add the German definite article in dative case to a preposition

Add the German definite article in dative case to a preposition

Usage

add_definite_article_de(
  preposition,
  gender = c("feminine", "masculine", "neuter")
)

Arguments

preposition

The preposition to add the German definite article in dative case to. A character scalar.

gender

The grammatical gender of the preposition.

Value

A character scalar.

See also