Decline a German noun

Note that this function currently doesn't support the special cases where a masculine noun gets an n or en ending attached in the accusative, dative and genitive cases.

decline_noun_de(
  noun,
  gender = c("feminine", "masculine", "neuter"),
  case = c("nominative", "genitive", "dative", "accusative")
)

Arguments

noun
The noun to decline. A character scalar.

gender
The grammatical gender of the noun.

case
The grammatical case of the noun.

Value

A character scalar.

See also